prereqs: Fix the script for Ubuntu 20
Python 2.x is considered to be so dead that the Ubuntu folks decided to
remove python-pip, and the only way of installing it is to curl|bash a
binary from the pip maintainers.

Rather than force our users to run this through the shell script, bail
out and warn the user.

Change-Id: Ie41fc2279ed62977d43e13b6ef0aba40cd7aa18e
GitOrigin-RevId: 67d7d423566c02c142f8473d81bc1c3f3f3ea7e0