renode: Add in a simple shell script to start things up
This allows for simple access to renode via ssh calls, without having to play
games with multiple ssh connections. This is basically just a timing ploy -- it
starts renode's repl in non-graphical mode and optionally starts includes a
script. Two seconds later in the other region, it starts up a telnet session on
renode's REPL port, 1234.

Change-Id: I172403c60cab24d476d722ffdfa53635026ee068
GitOrigin-RevId: be28721b33429661007e5fdbda8af7f408816703