downloads: Allow for the developer to prevent downloads
Not all developers need to check for the latest version of the IREE
toolchain, or the latest version of the Renode emulator to do their
day-to-day work, and keeping up to date with these tools often causes
significant downtime or delays in a build.

This change adjusts the scripts that do the downloading to honor the
PIN_TOOLCHAINS environment variable, allowing builds to continue
without the latest IREE or Renode toolchains.

PIN_TOOLCHAINS is expected to be a space-delimited list of toolchain
names.

Change-Id: I68ec58a3726e6b8130571e02b99d7b86b561b94d
GitOrigin-RevId: c8e0bcfcb63ec0923bc6bc06a21c474306b98577